Refactored project to new layout
