+++ Directory.Build.Props
@@ -1,0 +1,8 @@
+﻿<Project>
+  <PropertyGroup>
+    <AssemblyVersion>1.0.0.0</AssemblyVersion>
+    <FileVersion>1.0.0.0</FileVersion>
+    <InformationalVersion>1.0.0</InformationalVersion>
+    <Version>1.0.0</Version>
+  </PropertyGroup>
+</Project>
