+++ Paramore.Brighter.Extentions.sln
@@ -3,11 +3,18 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.26730.12
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Paramore.Brighter.AspNetCore", "src\Paramore.Brighter.AspNetCore\Paramore.Brighter.AspNetCore.csproj", "{44667A18-7967-4927-B2FA-5AD3907626C9}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Paramore.Brighter.Extensions.DependencyInjection", "src\Paramore.Brighter.Extensions.DependencyInjection\Paramore.Brighter.Extensions.DependencyInjection.csproj", "{44667A18-7967-4927-B2FA-5AD3907626C9}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Paramore.Brighter.HostedService", "Paramore.Brighter.HostedService\Paramore.Brighter.HostedService.csproj", "{C9141905-0607-4552-8086-C51A3F14FE0A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Paramore.Brighter.ServiceActivator.Extensions.Hosting", "src\Paramore.Brighter.ServiceActivator.Extensions.Hosting\Paramore.Brighter.ServiceActivator.Extensions.Hosting.csproj", "{C9141905-0607-4552-8086-C51A3F14FE0A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HostedServiceTest", "HostedServiceTest\HostedServiceTest.csproj", "{EB1B3B67-5164-46B7-9F9C-0473E43CC369}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HostedServiceTest", "HostedServiceTest\HostedServiceTest.csproj", "{EB1B3B67-5164-46B7-9F9C-0473E43CC369}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Paramore.Brighter.ServiceActivator.Extensions.DependencyInjection", "src\Paramore.Brighter.ServiceActivator.Extensions.DependencyInjection\Paramore.Brighter.ServiceActivator.Extensions.DependencyInjection.csproj", "{BB1D399D-FEC4-411A-A6DA-BA849F531A63}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{D765DE40-3626-47B8-915D-404E8D6A7F4B}"
+	ProjectSection(SolutionItems) = preProject
+		Directory.Build.Props = Directory.Build.Props
+	EndProjectSection
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,6 +34,10 @@ Global
 		{EB1B3B67-5164-46B7-9F9C-0473E43CC369}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EB1B3B67-5164-46B7-9F9C-0473E43CC369}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EB1B3B67-5164-46B7-9F9C-0473E43CC369}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BB1D399D-FEC4-411A-A6DA-BA849F531A63}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BB1D399D-FEC4-411A-A6DA-BA849F531A63}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BB1D399D-FEC4-411A-A6DA-BA849F531A63}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BB1D399D-FEC4-411A-A6DA-BA849F531A63}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
