--- Paramore.Brighter.HostedService/IServiceActivatorBuilder.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Reflection;
-using Paramore.Brighter.AspNetCore;
-
-namespace Paramore.Brighter.HostedService
-{
-    public interface IServiceActivatorBuilder : IBrighterHandlerBuilder
-    {
-        IServiceActivatorBuilder MapperRegistry(Action<HostMessageMapperRegistry> registerMappers);
-        IServiceActivatorBuilder MapperRegistryFromAssemblies(params Assembly[] assemblies);
-    }
-}
