--- Paramore.Brighter.HostedService/Paramore.Brighter.HostedService.csproj
@@ -1,16 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Hosting" Version="2.1.1" />
-    <PackageReference Include="Paramore.Brighter.ServiceActivator" Version="7.4.2" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\src\Paramore.Brighter.AspNetCore\Paramore.Brighter.AspNetCore.csproj" />
-  </ItemGroup>
-
-</Project>
