+++ src/Paramore.Brighter.Extensions.DependencyInjection/Paramore.Brighter.Extensions.DependencyInjection.csproj
@@ -1,15 +1,13 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <Version>1.0.0</Version>
     <TargetFrameworks>netstandard1.5;netstandard2.0</TargetFrameworks>
+    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
     <Authors>Daniel Stockhammer</Authors>
-    <Description>ASP.NET Core integration for Brighter</Description>
+    <Description>Brighter Microsoft.Extensions.DependencyInjection support</Description>
     <PackageProjectUrl>https://github.com/BrighterCommand/Brighter</PackageProjectUrl>
     <PackageLicenseUrl>https://raw.githubusercontent.com/BrighterCommand/Brighter/master/LICENCE.txt</PackageLicenseUrl>
     <RepositoryType>git</RepositoryType>
     <RepositoryUrl>git://github.com/BrighterCommand/Brighter</RepositoryUrl>
-    <GenerateAssemblyVersionAttribute>false</GenerateAssemblyVersionAttribute>
-    <GenerateAssemblyFileVersionAttribute>false</GenerateAssemblyFileVersionAttribute>
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="Paramore.Brighter" Version="7.4.2" />
@@ -18,6 +16,6 @@
     <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="1.1.1" />
   </ItemGroup>
   <ItemGroup Condition=" '$(TargetFramework)' == 'netstandard2.0' ">
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="2.1.0" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="2.1.1" />
   </ItemGroup>
 </Project>
